refactor(core): extract zeph-context crate from zeph-core monolith (phase 1)
Extracts context assembly, compaction, and summarization pure logic from
zeph-core into a new standalone zeph-context crate with no dependency on
zeph-core, breaking the compilation coupling between leaf crates and the
full agent loop.

New crate: crates/zeph-context
- assembler.rs: ContextAssembler, ContextAssemblyInput, ContextMemoryView,
  PreparedContext, memory_first_keep_tail, fetch_* helpers
- budget.rs: ContextBudget, BudgetAllocation (moved from context.rs)
- manager.rs: ContextManager, CompactionState, CompactionTier (moved from
  context_manager.rs)
- slot.rs: ContextSlot, CompactionOutcome, chunk_messages, cap_summary
- input.rs: ContextAssemblyInput, ContextMemoryView, IndexAccess trait
- summarization.rs: build_chunk_prompt, build_anchored_summary_prompt,
  build_metadata_summary, build_tool_pair_summary_prompt,
  remove_tool_responses_middle_out, extract_overflow_ref
- compression_feedback.rs: pure context-loss detection and classification
- microcompact.rs: low-value tool detection helpers
- error.rs: ContextError

zeph-core changes:
- impl Agent<C> methods delegate to zeph-context pure functions
- context/mod.rs: re-exports zeph-context types for backward compat
- Old duplicates (chunk_messages, cap_summary, ContextBudget,
  ContextManager, ContextAssembler) replaced with pub use zeph_context::*
- context_impls.rs: IndexAccess adapter for zeph-core's IndexState
- zeph-core LOC reduced by ~900 lines

Closes #2900 (phase 1 — zeph-commands extraction deferred to phase 2)
