+++ crates/zeph-core/src/agent/microcompact.rs
@@ -6,35 +6,17 @@
 //! Clears stale low-value tool output from context when the session has been
 //! idle longer than `gap_threshold_minutes`. This is a zero-LLM-cost in-memory
 //! operation that reduces context pressure before compaction runs.
+//!
+//! Pure helpers (`is_low_value_tool`, `find_preceding_tool_use_name`,
+//! `LOW_VALUE_TOOLS`, `CLEARED_SENTINEL_PREFIX`) live in
+//! [`zeph_context::microcompact`].
 
 use zeph_llm::provider::MessagePart;
 
 use crate::channel::Channel;
-
-/// Tool names whose output is considered low-value after a session gap.
-///
-/// Case-insensitive comparison is used at the call site.
-const LOW_VALUE_TOOLS: &[&str] = &[
-    "bash",
-    "shell",
-    "grep",
-    "rg",
-    "ripgrep",
-    "glob",
-    "find",
-    "web_fetch",
-    "fetch",
-    "web_search",
-    "search",
-    "read",
-    "cat",
-    "list_directory",
-];
-
-/// Sentinel content placed in cleared tool outputs.
-///
-/// Prefixed with `[cleared` so reload detection can skip already-cleared parts.
-const CLEARED_SENTINEL_PREFIX: &str = "[cleared";
+use zeph_context::microcompact::{
+    CLEARED_SENTINEL_PREFIX, find_preceding_tool_use_name, is_low_value_tool,
+};
 
 impl<C: Channel> super::Agent<C> {
     /// Returns a warning string when the prompt cache has likely expired due to session idle time.
@@ -95,12 +77,9 @@ impl<C: Channel> super::Agent<C> {
         let keep_recent = cfg.keep_recent;
         let messages = &mut self.msg.messages;
 
-        // Collect indices of compactable tool output parts across all messages.
-        // We track (msg_idx, part_idx) for ToolOutput, and msg_idx for ToolResult messages.
         let mut compactable: Vec<(usize, CompactTarget)> = Vec::new();
 
         for (msg_idx, msg) in messages.iter().enumerate() {
-            // Only assistant messages carry ToolOutput; user messages carry ToolResult.
             for (part_idx, part) in msg.parts.iter().enumerate() {
                 match part {
                     MessagePart::ToolOutput {
@@ -118,12 +97,9 @@ impl<C: Channel> super::Agent<C> {
                         compactable.push((msg_idx, CompactTarget::Output(part_idx)));
                     }
                     MessagePart::ToolResult { content, .. } => {
-                        // ToolResult has no compacted_at field — use sentinel detection.
                         if content.starts_with(CLEARED_SENTINEL_PREFIX) {
                             continue;
                         }
-                        // For ToolResult we need to know the associated tool name.
-                        // Walk back in the same message to find the preceding ToolUse part.
                         let tool_name = find_preceding_tool_use_name(&msg.parts, part_idx);
                         if let Some(name) = tool_name
                             && is_low_value_tool(name)
@@ -141,7 +117,6 @@ impl<C: Channel> super::Agent<C> {
             return;
         }
 
-        // Preserve the keep_recent most recent compactable entries.
         let clear_count = total.saturating_sub(keep_recent);
         if clear_count == 0 {
             tracing::debug!(
@@ -189,23 +164,6 @@ enum CompactTarget {
     Result(usize),
 }
 
-/// Returns the tool name from the closest preceding `ToolUse` part, if any.
-fn find_preceding_tool_use_name(parts: &[MessagePart], result_idx: usize) -> Option<&str> {
-    // Walk backward from result_idx - 1 looking for ToolUse.
-    for part in parts[..result_idx].iter().rev() {
-        if let MessagePart::ToolUse { name, .. } = part {
-            return Some(name.as_str());
-        }
-    }
-    None
-}
-
-/// Returns `true` if `tool_name` (case-insensitive) is in the low-value set.
-fn is_low_value_tool(tool_name: &str) -> bool {
-    let lower = tool_name.to_lowercase();
-    LOW_VALUE_TOOLS.contains(&lower.as_str())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -245,7 +203,6 @@ mod tests {
             gap_threshold_minutes: 0,
             keep_recent: 1,
         });
-        // last_assistant_at is None by default
         assert!(agent.cache_expiry_warning().is_none());
     }
 
@@ -256,7 +213,6 @@ mod tests {
             gap_threshold_minutes: 60,
             keep_recent: 1,
         });
-        // Set last_assistant_at to now — elapsed is ~0 minutes, well below 60.
         agent.session.last_assistant_at = Some(std::time::Instant::now());
         assert!(agent.cache_expiry_warning().is_none());
     }
@@ -268,52 +224,10 @@ mod tests {
             gap_threshold_minutes: 0,
             keep_recent: 1,
         });
-        // gap_threshold_minutes = 0 means any elapsed time exceeds the threshold.
         agent.session.last_assistant_at = Some(std::time::Instant::now());
         let warning = agent.cache_expiry_warning();
         assert!(warning.is_some());
         let msg = warning.unwrap();
         assert!(msg.contains("Cache expired"), "unexpected message: {msg}");
-    }
-
-    #[test]
-    fn low_value_tool_detection_case_insensitive() {
-        assert!(is_low_value_tool("Bash"));
-        assert!(is_low_value_tool("GREP"));
-        assert!(is_low_value_tool("list_directory"));
-        assert!(!is_low_value_tool("file_edit"));
-        assert!(!is_low_value_tool("memory_save"));
-        assert!(!is_low_value_tool("mcp_tool"));
-    }
-
-    #[test]
-    fn find_preceding_tool_use_name_returns_closest() {
-        use zeph_llm::provider::MessagePart;
-        let parts = vec![
-            MessagePart::ToolUse {
-                id: "1".into(),
-                name: "bash".into(),
-                input: serde_json::Value::Null,
-            },
-            MessagePart::ToolResult {
-                tool_use_id: "1".into(),
-                content: "output".into(),
-                is_error: false,
-            },
-        ];
-        let name = find_preceding_tool_use_name(&parts, 1);
-        assert_eq!(name, Some("bash"));
-    }
-
-    #[test]
-    fn find_preceding_tool_use_name_no_match() {
-        use zeph_llm::provider::MessagePart;
-        let parts = vec![MessagePart::ToolResult {
-            tool_use_id: "1".into(),
-            content: "output".into(),
-            is_error: false,
-        }];
-        let name = find_preceding_tool_use_name(&parts, 0);
-        assert!(name.is_none());
     }
 }
