+++ crates/zeph-core/src/agent/mod.rs
@@ -8,6 +8,7 @@ mod commands;
 pub(crate) mod compaction_strategy;
 pub(super) mod compression_feedback;
 mod context;
+mod context_impls;
 pub(crate) mod context_manager;
 mod corrections;
 pub mod error;
