+++ crates/zeph-core/src/agent/tool_execution/mod.rs
@@ -12,12 +12,6 @@ use crate::channel::Channel;
 use crate::redact::redact_secrets;
 use zeph_skills::loader::Skill;
 
-/// Prefix used in the overflow notice appended to tool outputs that exceed the size threshold.
-/// Shared with the pruning logic so both sides stay in sync if the format changes.
-///
-/// Current format: `[full output stored — ID: {uuid} — {bytes} bytes, use read_overflow tool to retrieve]`
-pub(crate) const OVERFLOW_NOTICE_PREFIX: &str = "[full output stored — ID: ";
-
 enum AnomalyOutcome {
     Success,
     Error,
